refactor: LottoResult (AS-IS) 제거, Controller, OutputView 에 WinningLotto, LottoResult 적용
