+++ src/test/java/lotto/domain/LottoResultTest.java
@@ -1,69 +1,45 @@
 package lotto.domain;
 
-import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 
-import java.util.List;
-import java.util.Map.Entry;
-import java.util.Set;
-import java.util.stream.Collectors;
-import java.util.stream.IntStream;
+import java.util.Map;
 import java.util.stream.Stream;
 
-import static lotto.domain.Rank.values;
+import static lotto.domain.Rank.FIFTH;
+import static lotto.domain.Rank.FIRST;
+import static lotto.domain.Rank.FOURTH;
+import static lotto.domain.Rank.NONE;
+import static lotto.domain.Rank.SECOND;
+import static lotto.domain.Rank.THIRD;
 import static org.assertj.core.api.Assertions.assertThat;
 
 class LottoResultTest {
 
-    @Test
-    void 결과에는_모든_등수에_대한_당첨갯수가_있어야_한다() {
-        LottoNumbers winner = new LottoNumbers("1,2,3,4,5,6");
-        LottoNumber bonusLottoNumber = LottoNumber.valueOf(7);
-        Games games = new Games(List.of(winner, winner));
-
-        LottoResult lottoResult = new LottoResult(winner, bonusLottoNumber, games);
-        Set<Entry<Rank, Integer>> entrySet = lottoResult.countPerRank().entrySet();
-
-        assertThat(entrySet).hasSize(values().length);
-    }
-
     @ParameterizedTest
-    @MethodSource("gamesAndEarningRate")
-    void 전체_게임에_대한_수익률을_구할_수_있다(Games games, double expected) {
-        LottoNumbers winner = new LottoNumbers("1,2,3,4,5,6");
-        LottoNumber bonusLottoNumber = LottoNumber.valueOf(7);
-        LottoResult lottoResult = new LottoResult(winner, bonusLottoNumber, games);
+    @MethodSource("rankAndFrequency")
+    void 등수의_빈도를_반환한다(Rank rank, Integer expected) {
+        LottoResult lottoResult = new LottoResult(Map.of(
+                FIRST, 1,
+                SECOND, 2,
+                THIRD, 3,
+                FOURTH, 4,
+                FIFTH, 5,
+                NONE, 6
+        ), 0);
 
-        assertThat(lottoResult.calculateEarningRate()).isEqualTo(expected);
+        assertThat(lottoResult.of(rank)).isEqualTo(expected);
     }
 
-    private static Stream<Arguments> gamesAndEarningRate() {
-        double twice = 2.00;
-        double breakEven = 1.00;
-        double twoThird = 0.67;
-        double half = 0.50;
-        double oneThird = 0.33;
-
+    private static Stream<Arguments> rankAndFrequency() {
         return Stream.of(
-                Arguments.arguments(new Games(getFourthAndNoneGame(2, 3)), twice),
-                Arguments.arguments(new Games(getFourthAndNoneGame(1, 4)), breakEven),
-                Arguments.arguments(new Games(getFourthAndNoneGame(2, 13)), twoThird),
-                Arguments.arguments(new Games(getFourthAndNoneGame(1, 9)), half),
-                Arguments.arguments(new Games(getFourthAndNoneGame(1, 14)), oneThird)
+                Arguments.arguments(FIRST, 1),
+                Arguments.arguments(SECOND, 2),
+                Arguments.arguments(THIRD, 3),
+                Arguments.arguments(FOURTH, 4),
+                Arguments.arguments(FIFTH, 5),
+                Arguments.arguments(NONE, 6)
         );
-    }
-
-    private static List<LottoNumbers> getFourthAndNoneGame(int fourth, int none) {
-        List<LottoNumbers> fourthGames = IntStream.range(0, fourth)
-                .mapToObj(i -> new LottoNumbers("1,2,3,7,8,9"))
-                .collect(Collectors.toList());
-        List<LottoNumbers> noneGames = IntStream.range(0, none)
-                .mapToObj(i -> new LottoNumbers("1,2,7,8,9,10"))
-                .collect(Collectors.toList());
-        fourthGames.addAll(noneGames);
-
-        return fourthGames;
     }
 }
